renamed to filter.js
