+++ src/store/reducers/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
 import { booksReducer } from './booksReducer';
-import filterReducer from './filterReducer';
+import filterReducer from './filter';
 
 export default combineReducers({ booksReducer, filterReducer });
